requirements.txt
# Core dependencies
google-adk>=0.3.0
litellm>=1.35.0
groq>=0.5.0

# Google Gemini (for image analysis)
google-generativeai>=0.8.0

# Image processing
Pillow>=10.0.0

# Async database
asyncpg>=0.29.0

# Redis
redis>=5.0.0

# HTTP client
httpx>=0.27.0

# Retry logic
tenacity>=8.2.0

# Environment configuration
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.9.0
